+++ lib/src/lib.rs
@@ -31,7 +31,6 @@ pub mod git_backend;
 pub mod gitignore;
 pub mod hex_util;
 pub mod index;
-pub mod index_store;
 #[cfg(feature = "legacy-thrift")]
 mod legacy_thrift_op_store;
 pub mod local_backend;
